5/23/2024 fcm added but milestone start date and end date are issue

+++ src/main/java/com/task_pay/task_pay/models/entities/Task.java
@@ -20,7 +20,7 @@ public class Task {
     private String taskName;
 
     @Column(nullable = false)
-    private Integer taskPrice;
+    private double taskPrice;
 
     @Column(nullable = false)
     private String taskAbout;
